Fix 42883: "date + bigint" existiert nicht, Tage-Summe auf int casten
Fehler beim Ausfuehren der Migration: "operator does not exist: date +
bigint". Ursache: sum() liefert von sich aus bigint, und
tage_vor_letztem_abschnitt war entsprechend als bigint gecastet - Postgres
kennt aber nur "date + integer", kein "date + bigint".

Fix: sum(tage) auf ::int statt ::bigint casten (Tage passen locker in
int), damit die Austrittsdatum-Berechnung "beginn_letzter_abschnitt +
(105 - tage_vor_letztem_abschnitt - vorbeschaeftigung) - 1" wieder
aufgeht. Betrifft schema.sql und die Migration gleichermassen.

+++ supabase/migration_2026-08-11_15_wochen_zusammenrechnung.sql
@@ -171,12 +171,15 @@ w as (
     saison_jahr,
     min(von) as erster_arbeitstag,
     max(bis) as letzter_arbeitstag,
-    sum(tage)::bigint as beschaeftigungstage,
+    -- ::int (nicht bigint): tage_vor_letztem_abschnitt unten wird auf ein
+    -- Datum addiert, und Postgres kennt nur "date + integer", kein
+    -- "date + bigint" (Fehler 42883). sum() liefert von sich aus bigint.
+    sum(tage)::int as beschaeftigungstage,
     count(*)::int as anzahl_abschnitte,
     -- Für das Austrittsdatum: der laufende (zuletzt begonnene) Abschnitt
     -- darf noch so viele Kalendertage laufen, wie vom Budget übrig ist.
     (array_agg(von order by von desc))[1] as beginn_letzter_abschnitt,
-    (sum(tage) - (array_agg(tage order by von desc))[1])::bigint
+    (sum(tage) - (array_agg(tage order by von desc))[1])::int
       as tage_vor_letztem_abschnitt
   from abschnitte
   group by employee_id, saison_jahr
